changed tests to pytest
